requirements.txt
google-genai>=1.30.0
aiosqlite>=0.21.0
python-dotenv>=1.0.0